# openapi/v1.yaml
...
openapi: 3.1.0
info:
  title: cafaye guard
  # This is a new document, so it starts at 1.0.0 and there is no compatibility
  # promise to keep. It is the document's own semantic version and it follows
  # guard's release line, not the `/v1` prefix: the two answer different
  # questions, and core's sync rule asks for both.
  #
  # Non-breaking additions bump this and nothing else. A breaking change bumps the
  # `/v1` prefix and this in the same commit, and `/v1` is never mutated in place.
  version: 1.0.0
  summary: The cafaye public gateway — authentication, rate limits, and the browser sign-in surface.
  description: |
    `guard` is the only door into a cafaye deployment. It authenticates callers,
    limits what they can ask for, and — once the routing packet lands — forwards
    requests to the service that owns the work. It holds no business data and
    publishes no events.

    ## Two surfaces, two credentials, and they do not cross

    **`/v1/*` is the API surface.** A bearer JWT verified locally against
    `identity`'s published JWKS, or an API key. Never a cookie. A request with
    neither is `401` before any handler runs, and the `/v1/*` mount is what makes
    that true for every route on the prefix rather than for the ones somebody
    remembered to extend.

    **`/auth/*` is the browser surface.** A session id in a `__Host-bff-session`
    cookie that guard minted, swapped for the identity token server-side when a
...
    provably same-origin.

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/me:
    get:
      tags: [api]
      operationId: getCurrentPrincipal
      summary: The verified principal
      description: |
        Resolves the presented credential and returns what guard proved about the
        caller: the `sub`, the scopes, the account, and the claims themselves.

        This is the operation that exists to prove the auth chain is wired end to
        end. It has no behaviour of its own and it forwards nothing; the routing
        packet replaces it with real traffic.

        A missing, malformed, expired, wrongly-signed, wrongly-issued or
        wrongly-addressed token, and an API key that is unknown or withdrawn, are
        all **one** `401` with one body. Which of the four it was is not something
        a caller learns by trying.

        **A `503`, never a `401`, when `identity` is unreachable.** The signing
        keys could not be fetched, so guard cannot tell whether the token is good;
        a `401` would send a caller to rotate a credential that was fine.
      security:
        - bearerToken: []
        - apiKeyCredential: []
      responses:
        '200':
          description: The verified principal.
          headers:
            RateLimit-Policy:
              $ref: '#/components/headers/RateLimitPolicy'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
            X-RateLimit-Limit:
              $ref: '#/components/headers/XRateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/XRateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/XRateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
              example:
                sub: ab000000-0000-0000-0000-000000000000
                scope: [invoices:write]
                accountId: cd000000-0000-0000-0000-000000000000
                claims:
                  iss: https://identity.localhost
                  aud: guard
                  sub: ab000000-0000-0000-0000-000000000000
                  account_id: cd000000-0000-0000-0000-000000000000
                  scope: invoices:write
                  exp: 1790880000
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/Throttled'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /auth/register:
    post:
      tags: [browser]
      operationId: registerUser
      summary: Register an account
      description: |
        Creates an account through `identity` and returns the two fields a
        browser needs. **No session is created**: a registration is not a login,
        and the caller signs in separately with `POST /auth/login`.

        The body is forwarded to `identity` **verbatim**. guard decides only
        whether it is a JSON object and whether it is under 4 KiB; whether the
        address is deliverable and the password strong enough is `identity`'s
        call, under the same rules every other client of `identity` gets. A second
        set of rules in the gateway would be a second answer to "what is a valid
        password".

        This is the only operation in the gateway that reveals whether an address
        is already registered, and it answers `409`. Everywhere else, "no such
        account" and "wrong password" are indistinguishable.

        Requires a provably same-origin request: `Sec-Fetch-Site: same-origin`, or
        an `Origin` header naming this host. See open decision 4 in the header.
      security: []
      parameters:
        - $ref: '#/components/parameters/SecFetchSite'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: kaka@example.com
              password: correct horse battery staple
      responses:
        '201':
          description: The account was created. No cookie is set.
          headers:
            RateLimit-Policy:
              $ref: '#/components/headers/RateLimitPolicy'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
            X-RateLimit-Limit:
              $ref: '#/components/headers/XRateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/XRateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/XRateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: ab000000-0000-0000-0000-000000000000
                email: kaka@example.com
        '400':
          $ref: '#/components/responses/MalformedBody'
        '403':
          $ref: '#/components/responses/NotSameOrigin'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/BodyTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/Throttled'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /auth/login:
    post:
      tags: [browser]
      operationId: createSession
      summary: Log in
      description: |
        Exchanges an email and a password for a session id in a
        `__Host-bff-session` cookie.

        **The response body carries `expires_at` and nothing else.** The identity
        token is never echoed to the page: a token in a page is a token a
        cross-site script can read, which is the entire reason this surface
        exists. The cookie is `Secure`, `HttpOnly`, `SameSite=Lax` and `Path=/`,
        with no `Domain`; the `__Host-` prefix is a contract the browser
        enforces, and it is what stops a sibling or subdomain origin from setting
        or overwriting the cookie.

        The session id is minted here and is never accepted from the caller.
        Session fixation needs an id the attacker knows first.

        A wrong password and an unregistered address are **one** `401` with one
        body, and comparable time. A locked account is `423` with `Retry-After`,
        and the window travels with the answer — collapsing it into the `401`
        would tell somebody with the right password to try again and be refused
        again for a quarter of an hour.

        Requires a provably same-origin request. See open decision 4 in the
        header.
      security: []
      parameters:
        - $ref: '#/components/parameters/SecFetchSite'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: kaka@example.com
              password: correct horse battery staple
      responses:
        '200':
          description: |
            The credentials were correct and a session now exists. The body says
            when it runs out; the cookie is how the browser presents it.
          headers:
            Set-Cookie:
              description: The `__Host-bff-session` cookie, or absent for a client that ignores cookies.
              schema:
                type: string
              example: '__Host-bff-session=…; Path=/; Secure; HttpOnly; SameSite=Lax'
            RateLimit-Policy:
              $ref: '#/components/headers/RateLimitPolicy'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
            X-RateLimit-Limit:
              $ref: '#/components/headers/XRateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/XRateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/XRateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
              example:
                expires_at: '2026-09-30T12:00:00Z'
        '400':
          $ref: '#/components/responses/MalformedBody'
        '401':
          description: |
            The credentials were refused. The body is identical whether the
            address is unregistered or the password is wrong, and says nothing
            about which.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://errors.cafaye.com/unauthorized
                title: Unauthorized
                status: 401
                detail: email or password is not correct
                instance: /auth/login
                code: unauthorized
                trace_id: 0af7651916cd43dd8448eb211c80319c
        '403':
          $ref: '#/components/responses/NotSameOrigin'
        '413':
          $ref: '#/components/responses/BodyTooLarge'
        '423':
          $ref: '#/components/responses/AccountLocked'
        '429':
          $ref: '#/components/responses/Throttled'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /auth/logout:
    post:
      tags: [browser]
      operationId: deleteSession
      summary: Log out
      description: |
        Revokes the session the presented cookie names, and clears it. **The
        local session is dropped first, on purpose:** failing to sign a browser
        out is worse than a token that outlives its revocation until its `exp`.

        If `identity` cannot be told, this is still a `204` and the cookie is
        still cleared — the local session is gone and the browser is signed out of
        this gateway. The `503` is only reached when the local drop itself did not
        happen, and the detail says the revocation is unconfirmed rather than
        leaving the caller to assume it.

        A cookie the store has forgotten is `204`, not `401`: the browser is not
        signed in, there is provably nothing to revoke, and refusing is what makes
        sign-out look broken.

        Requires a provably same-origin request.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/SecFetchSite'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: The session is gone and the cookie is cleared. No body.
          headers:
            Set-Cookie:
              description: An expired `__Host-bff-session` cookie, repeating every attribute the setting cookie had.
              schema:
                type: string
              example: '__Host-bff-session=; Path=/; Secure; HttpOnly; SameSite=Lax; Max-Age=0'
            RateLimit-Policy:
              $ref: '#/components/headers/RateLimitPolicy'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
            X-RateLimit-Limit:
              $ref: '#/components/headers/XRateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/XRateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/XRateLimitReset'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/NotSameOrigin'
        '429':
          $ref: '#/components/responses/Throttled'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The API surface. An RS256 JWT issued by `identity` and verified locally
        against its published JWKS, cached by `kid` with a bounded TTL.

        **RS256 only.** `alg: none`, HS256 and anything the key set does not
        advertise are refused — and refused on the *protected header*, before any
        key is fetched, so a forged token costs no network call and never aims
        traffic at `identity`. The accepted algorithm is one constant in
        `src/middleware/jwt.ts`, never something a token can widen. core's
        conventions also allow ES256; widening is a one-line change and belongs
        with the moment `identity` publishes ES256 keys.

        Required claims: `iss` (the configured issuer), `aud` (guard's own client
        id), `sub`, and `exp`. The capability claim is a space-separated `scope`
        string — not core's `scopes` — and an absent `scope` is an empty set,
        never everything. Both are open in `cafaye.yml`.
    apiKeyCredential:
      type: apiKey
      in: header
      name: Authorization
      description: |
        A machine credential for an account, presented as
        `Authorization: ApiKey caf_<43 base64url characters>`. Declared as an
        `apiKey` in the `Authorization` header rather than as an `http` scheme
        because `ApiKey` is not one of RFC 7235's registered schemes.

        32 bytes from the platform CSPRNG. Only its SHA-256 is stored, only its
        first twelve characters are ever displayed, and it is shown exactly once —
        at issue. A revocation takes effect on the very next request, and an
        unknown key and a withdrawn key are one refusal, because telling them
        apart would confirm that a guessed key once existed.

        The scopes on a key are in the same vocabulary as a token's `scope`, and
        the key sets the same principal: everything downstream treats the two
        credentials as the same caller.

...